Remove the classes from grid.py and solving.py that don't belong there.
This removes the classes in those files that don't belong, and renames
the main function to be the Solver class, which is re-written in a
followup commit.
